Move CaughtFatalException to new exceptions module
