+++ pygametemplate/exceptions.py
